// angularProject/src/app/api/model/authenticateResponse.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthenticateResponse { 
    email?: string | null;
    phoneNumber?: string | null;
    customerId?: number;
    token?: string | null;
    isAdvisor?: boolean;
    customerName?: string | null;
    is18OrAbove?: boolean;
    isWhatsAppNumber?: boolean;
    isAdmin?: boolean;
    isEmployee?: boolean;
    customerRole?: number;
    profileImage?: string | null;
    companyId?: number;
    userName?: string | null;
    isPrimeCandidate?: boolean;
    isPrimeRecruiter?: boolean;
    isContractFreelancer?: boolean;
    countryId?: string | null;
    affiliateId?: number;
}