// angularProject/src/app/api/model/loginCustomerRequest.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginCustomerRequest { 
    phoneNumber?: string | null;
    email?: string | null;
    password?: string | null;
}